requirements.txt
click==7.1.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==1.4.2
python-dotenv==0.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
